This PR contains more optimizations apart from the ones in:
- https://github.com/gballet/go-verkle/pull/314
- https://github.com/crate-crypto/go-ipa/pull/31

The `go.mod` in this PR uses those branches; we should update the official commits whenever those PRs are merged.

Apart from that, we're doing other optimizations that will be explained in PR comments.

The results are significant comparing `beverly-hills` vs this PR:
```
name                               old time/op    new time/op    delta
TriesRandom/VKT/1000_accounts-16     56.8ms ± 3%    22.9ms ± 3%  -59.73%  (p=0.001 n=10+5)
TriesRandom/VKT/5000_accounts-16      264ms ± 2%      76ms ± 2%  -71.34%  (p=0.001 n=10+5)
TriesRandom/VKT/10000_accounts-16     535ms ± 2%     146ms ± 5%  -72.78%  (p=0.001 n=10+5)

name                               old alloc/op   new alloc/op   delta
TriesRandom/VKT/1000_accounts-16     9.56MB ± 0%    9.76MB ± 0%   +2.13%  (p=0.001 n=9+5)
TriesRandom/VKT/5000_accounts-16     43.5MB ± 0%    43.9MB ± 0%   +0.94%  (p=0.002 n=8+5)
TriesRandom/VKT/10000_accounts-16    87.5MB ± 0%    88.2MB ± 0%   +0.83%  (p=0.001 n=10+5)

name                               old allocs/op  new allocs/op  delta
TriesRandom/VKT/1000_accounts-16      32.8k ± 0%     35.5k ± 0%   +8.08%  (p=0.001 n=9+5)
TriesRandom/VKT/5000_accounts-16       151k ± 0%      144k ± 0%   -4.88%  (p=0.001 n=10+5)
TriesRandom/VKT/10000_accounts-16      303k ± 0%      285k ± 0%   -5.97%  (p=0.001 n=10+5)
```