@jsign 2023-03-23
Should be interesting if we consider https://github.com/gballet/go-verkle/pull/314, but for now I'll close this PR.
We can recover it if we go that route.